server/pyproject.toml
[project]
name = "ai-research-agent-mcp"
version = "0.1.0"
description = "MCP-powered AI Research Engineer with autonomous research, coding, and reporting capabilities"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "mcp>=1.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "duckduckgo-search>=4.0.0",
    "trafilatura>=1.6.0",
    "httpx>=0.25.0",
    "chromadb>=0.4.0",
    "anthropic>=0.18.0",
    "sentence-transformers>=2.2.0",
    "pypdf>=3.17.0",
    "python-docx>=1.1.0",
    "RestrictedPython>=6.2",
    "asteval>=0.9.31",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "aiofiles>=23.0.0",
    "tenacity>=8.2.0",
    "structlog>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "ruff>=0.0.290",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py310"
